proof of concept - raw FullScreenshot
As for now, the function might work as checking the full clientHeight,
parsing it as an integer, setting the viewport (might be a good idea to
lose hardcoded "1920" in favour of clientWidth or a flag making in HD,
don't know yet) and then getting the screnshot and saving it.

Few issues with this solution:
* Maybe there's a better way of evaluating the JavaScript
* Maybe there's no need to evaluate it whatsoever
* What happens with the viewport after the screenshot has been made?
  Right now it will be morphed to a full height of the page.

Will keep working on it as I have to explore the library a bit more.