remove the dependency gorilla/websocket
Since Chrome only uses a subset of the WebSocket protocol, and current WebSocket libs on the market don't support dynamic-sized frames, so I think it's better to implement a customized client for chrome.

Because of the current limitation, each cdp client will consume some amount of fix-sized memory, such as the default value is 2MB. After this ticket, the memory overhead of the cdp client will be zero which will make controlling hundreds of Chromes by a single machine way less memory demanding.

resolve #248